made result id field an int
Former-commit-id: 4bd4f96bd2cff84ad93749e33f8d6851ee50e410

+++ src/mitll/langtest/client/LangTestDatabaseAsync.java
@@ -136,7 +136,7 @@ public interface LangTestDatabaseAsync {
                              int userid, String flText, String which,
                              AsyncCallback<Collection<String>> async);
 
-  void addRoundTrip(long resultid, int roundTrip, AsyncCallback<Void> async);
+  void addRoundTrip(int resultid, int roundTrip, AsyncCallback<Void> async);
 
   void getResultASRInfo(int resultID, int width, int height, AsyncCallback<PretestScore> async);
